# 🚀 Weekend 2 Epic: Universal Memory Optimization with 6.5x GPU Speedup & 196.7 GFLOPS CPU