--- a/Makefile.smart
+++ b/Makefile.smart
@@ -68,6 +68,7 @@ COMMON_MODULES = \
     $(SRC_DIR)/sparkle_gpu_kernels.f90 \
     $(SRC_DIR)/sparkle_glsl_generator.f90 \
     $(SRC_DIR)/sparkle_rdna_shader_generator.f90 \
+    $(SRC_DIR)/sparkle_dynamic_shader_system.f90 \
     $(SRC_DIR)/sparkle_adaptive_kernel.f90 \
     $(SRC_DIR)/sparkle_kernel_variants.f90 \
     $(SRC_DIR)/sparkle_shader_parser.f90 \
@@ -417,6 +418,20 @@ $(BUILD_DIR)/test_rdna_shader_optimization: $(OBJECTS) $(C_OBJECTS) $(EXAMPLES_D
 	@echo "🔨 Building RDNA shader optimization test..."
 	$(FC) $(BASE_FFLAGS) -I$(BUILD_DIR) $(OBJECTS) $(C_OBJECTS) \
 		$(EXAMPLES_DIR)/test_rdna_shader_optimization.f90 -o $@ $(LDFLAGS)
+
+# Dynamic shader system test
+test_dynamic_shader_system: $(BUILD_DIR)/test_dynamic_shader_system
+	@echo "🚀 Testing dynamic shader system..."
+	@./$(BUILD_DIR)/test_dynamic_shader_system
+
+$(BUILD_DIR)/test_dynamic_shader_system: $(BUILD_DIR)/sparkle_types.o $(BUILD_DIR)/sparkle_glsl_generator.o $(BUILD_DIR)/sparkle_rdna_shader_generator.o $(BUILD_DIR)/sparkle_dynamic_shader_system.o $(EXAMPLES_DIR)/test_dynamic_shader_system.f90
+	@echo "🔨 Building dynamic shader system test..."
+	$(FC) $(BASE_FFLAGS) -I$(BUILD_DIR) \
+		$(BUILD_DIR)/sparkle_types.o \
+		$(BUILD_DIR)/sparkle_glsl_generator.o \
+		$(BUILD_DIR)/sparkle_rdna_shader_generator.o \
+		$(BUILD_DIR)/sparkle_dynamic_shader_system.o \
+		$(EXAMPLES_DIR)/test_dynamic_shader_system.f90 -o $@
 
 # Memory wall simple test
 test_memory_wall_simple: $(BUILD_DIR)/test_memory_wall_simple