… BENCHMARKS.md
… CLAUDE.md
… README.md
… docs/AMDGPU_DIRECT_INTEGRATION.md
… docs/CURRENT_STATE.md
… docs/DYNAMIC_SHADER_SYSTEM_IMPLEMENTED.md
… docs/GPU_ASYNC_BREAKTHROUGH.md
… docs/GPU_ASYNC_PERFORMANCE_ANALYSIS.md
… docs/GPU_ASYNC_REALITY_CHECK.md
… docs/GPU_IDLE_TIME_INSIGHT.md
… docs/GPU_UNIQUE_FEATURES.md
… docs/RDNA_ARCHITECTURAL_OPTIMIZATION.md
… docs/UNIVERSAL_DEVICE_SELECTION.md
… docs/UNIVERSAL_MEMORY_OPTIMIZATION_BREAKTHROUGH.md
… docs/Weekend2.md
… examples/test_amdgpu_direct_integration.f90
… examples/test_dynamic_shader_system.f90
… examples/test_gpu_async_executor.f90
… examples/test_gpu_async_honest.f90
… examples/test_gpu_async_poc.f90
… examples/test_gpu_dynamic_shaders.f90
… examples/test_gpu_shader_variants.f90
… examples/test_intelligent_device_juggling.f90
… examples/test_memory_wall_breakthrough.f90
… examples/test_memory_wall_simple.f90
… examples/test_peak_cpu_performance.f90
… examples/test_rdna_shader_optimization.f90
… examples/test_simd_performance.f90
… examples/test_universal_memory_optimization.f90
… src/amdgpu_device.f90